fix(gfql): the binding-name alias decline also guards the deprecated chain() surface
The typed E108 decline for an alias named like a same-frame binding column lived only on the gfql() dispatch; g.chain() still reached the engines and died with a raw pandas ValueError or polars DuplicateError. The helper moves to chain.py and runs on both surfaces; pins cover both surfaces (decline) and the cross-frame case (served, marker present).
