+++ graphistry/compute/chain.py
@@ -1027,6 +1027,54 @@ def _try_chain_fast_path(
         g.nodes(nodes).edges(edges), alias_n0, alias_e1, alias_n2, src, dst, node, direction)
 
 
+def reject_alias_named_like_binding(
+    g: Plottable, chain_obj: "Chain", *, include_edge_endpoint_aliases: bool = False
+) -> None:
+    """Typed decline for an alias named after a binding column: a node alias equal to the
+    node-ID binding, and (native chains and Cypher alike) an edge alias equal to the source,
+    destination or edge-ID binding.
+
+    The alias marker is stamped as ``<alias> = True``, so an alias equal to the node-id
+    column overwrites the ids themselves: pandas then died with a raw
+    ``ValueError: The column label 'id' is not unique`` from the chain's own merge while
+    polars answered ``True``. Neither is a usable result; decline the same way on both.
+    """
+    from graphistry.compute.exceptions import ErrorCode, GFQLValidationError
+    node_id = getattr(g, "_node", None)
+    endpoint_cols = {
+        col for col in (getattr(g, "_source", None), getattr(g, "_destination", None), getattr(g, "_edge", None))
+        if isinstance(col, str)
+    }
+    for op in chain_obj.chain:
+        if isinstance(node_id, str) and isinstance(op, ASTNode) and getattr(op, "_name", None) == node_id:
+            raise GFQLValidationError(
+                ErrorCode.E108,
+                "A node alias cannot be named after the node-ID binding column",
+                field="chain.name",
+                value=node_id,
+                suggestion=(
+                    f"The alias flag is materialized as a column named '{node_id}', which would "
+                    f"overwrite the node-ID binding. Rename the alias."
+                ),
+            )
+        if (
+            include_edge_endpoint_aliases
+            and isinstance(op, ASTEdge)
+            and getattr(op, "_name", None) in endpoint_cols
+        ):
+            raise GFQLValidationError(
+                ErrorCode.E108,
+                "An edge alias cannot be named after an edge endpoint binding column",
+                field="chain.name",
+                value=getattr(op, "_name", None),
+                suggestion=(
+                    "The alias flag is materialized as a column named like the edge "
+                    "source, destination or edge-ID binding, which would overwrite it. "
+                    "Rename the alias."
+                ),
+            )
+
+
 @otel_traced("gfql.chain", attrs_fn=_chain_otel_attrs)
 def chain(
     self: Plottable,
@@ -1088,6 +1136,7 @@ def _chain_with_strictness(
     # _coerce_input_formats then converts input formats (polars, arrow, spark, dask) to that engine.
     if isinstance(engine, str):
         engine = EngineAbstract(engine)
+    reject_alias_named_like_binding(self, ops if isinstance(ops, Chain) else Chain(ops), include_edge_endpoint_aliases=True)
     from graphistry.compute.ComputeMixin import _coerce_input_formats  # lazy — avoids circular import
     engine_concrete_early = resolve_engine(engine, self)
     if engine_concrete_early in (Engine.POLARS, Engine.POLARS_GPU):
